refactor: e2e_support wave-3 Slice C shared hosts and densify >400 suites
Extract cloned widget-test hosts into test/support libraries and densify
the seven suites that exceeded 400 lines so AC5–AC6 can land before the
lib/test ratchets in Slice D.
