+++ packages/colonizethis_app_e2e_support/test/e2e_wait_for_next_turn_label_advance_test.dart
@@ -10,171 +10,29 @@
 /// § CI), so the widget-test layer carries the behavioral pins.
 library;
 
-import 'dart:async';
-
-import 'package:colonizethis_app/features/game/screens/game/game_screen_shared.dart';
 import 'package:colonizethis_app/features/game/flame/overlays/turn_resolution_processing_dialog.dart';
-import 'package:colonizethis_app_l10n/l10n/l10n.dart';
+import 'package:colonizethis_app/features/game/screens/game/game_screen_shared.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
 
-/// Host that mounts a key-tagged next-turn label (and an optional
-/// [TurnResolutionProcessingDialog]) and exposes a fake-async `Timer`-driven
-/// hook for tests to flip the label or dismiss the dialog while the
-/// helper's `tester.pump` loop is running.
-///
-/// `Timer` callbacks scheduled in [State.initState] fire when `tester.pump`
-/// advances fake-time past the registered duration, so the helper observes
-/// the flip on a later polling iteration without the test needing to call
-/// `tester.pump` itself (which would deadlock against the helper's guarded
-/// pump).
-class _NextTurnLabelHost extends StatefulWidget {
-  const _NextTurnLabelHost({
-    required this.controller,
-    this.flipAfter,
-    this.flipToLabel,
-    this.flipToShowDialog,
-  });
-
-  final _NextTurnLabelController controller;
-
-  /// Fake-async delay before the host flips the controller, or `null` to
-  /// leave the controller untouched (no scheduled flip).
-  final Duration? flipAfter;
-
-  /// New label to apply when [flipAfter] elapses, or `null` to leave the
-  /// label unchanged.
-  final String? flipToLabel;
-
-  /// New processing-dialog visibility to apply when [flipAfter] elapses,
-  /// or `null` to leave the visibility unchanged.
-  final bool? flipToShowDialog;
-
-  @override
-  State<_NextTurnLabelHost> createState() => _NextTurnLabelHostState();
-}
-
-class _NextTurnLabelHostState extends State<_NextTurnLabelHost> {
-  Timer? _flipTimer;
-
-  @override
-  void initState() {
-    super.initState();
-    widget.controller.addListener(_onControllerChanged);
-    final after = widget.flipAfter;
-    if (after != null) {
-      _flipTimer = Timer(after, _applyFlip);
-    }
-  }
-
-  void _applyFlip() {
-    final newLabel = widget.flipToLabel;
-    if (newLabel != null) {
-      widget.controller.label = newLabel;
-    }
-    final newDialog = widget.flipToShowDialog;
-    if (newDialog != null) {
-      widget.controller.showProcessingDialog = newDialog;
-    }
-  }
-
-  @override
-  void dispose() {
-    _flipTimer?.cancel();
-    widget.controller.removeListener(_onControllerChanged);
-    super.dispose();
-  }
-
-  void _onControllerChanged() {
-    if (!mounted) return;
-    setState(() {});
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final showDialog = widget.controller.showProcessingDialog;
-    final label = widget.controller.label;
-    return Stack(
-      children: [
-        Center(
-          child: TextButton(
-            key: kGameMapNextTurnButtonKey,
-            onPressed: () {},
-            child: Text(label),
-          ),
-        ),
-        if (showDialog)
-          const Positioned.fill(
-            child: ColoredBox(
-              color: Color(0x88000000),
-              child: Center(
-                child: TurnResolutionProcessingDialog(
-                  phaseText: 'Validating orders...',
-                ),
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-class _NextTurnLabelController extends ChangeNotifier {
-  _NextTurnLabelController({
-    required String initialLabel,
-    bool initialShowProcessingDialog = false,
-  })  : _label = initialLabel,
-        _showProcessingDialog = initialShowProcessingDialog;
-
-  String _label;
-  bool _showProcessingDialog;
-
-  String get label => _label;
-
-  set label(String value) {
-    if (_label == value) {
-      return;
-    }
-    _label = value;
-    notifyListeners();
-  }
-
-  bool get showProcessingDialog => _showProcessingDialog;
-
-  set showProcessingDialog(bool value) {
-    if (_showProcessingDialog == value) {
-      return;
-    }
-    _showProcessingDialog = value;
-    notifyListeners();
-  }
-}
+import 'support/next_turn_label_harness.dart';
 
 Future<void> _pumpHost(
   WidgetTester tester,
-  _NextTurnLabelController controller, {
+  NextTurnLabelController controller, {
   Duration? flipAfter,
   String? flipToLabel,
   bool? flipToShowDialog,
-}) async {
-  await tester.pumpWidget(
-    MaterialApp(
-      localizationsDelegates: AppLocalizationsBinding.localizationsDelegates,
-      supportedLocales: AppLocalizations.supportedLocales,
-      home: Scaffold(
-        body: _NextTurnLabelHost(
-          controller: controller,
-          flipAfter: flipAfter,
-          flipToLabel: flipToLabel,
-          flipToShowDialog: flipToShowDialog,
-        ),
-      ),
-    ),
-  );
-}
+}) => pumpNextTurnLabelHost(
+  tester,
+  controller,
+  flipAfter: flipAfter,
+  flipToLabel: flipToLabel,
+  flipToShowDialog: flipToShowDialog,
+);
 
 void main() {
   suppressLogsForTests();
@@ -182,7 +40,7 @@ void main() {
   testWidgets(
     'returns immediately when current label already differs from before',
     (WidgetTester tester) async {
-      final controller = _NextTurnLabelController(
+      final controller = NextTurnLabelController(
         initialLabel: 'Next turn (2 / 1492)',
       );
       await _pumpHost(tester, controller);
@@ -213,7 +71,7 @@ void main() {
   testWidgets('returns once a scheduled label flip lands during pump', (
     WidgetTester tester,
   ) async {
-    final controller = _NextTurnLabelController(
+    final controller = NextTurnLabelController(
       initialLabel: 'Next turn (1 / 1492)',
     );
     // Schedule the host to flip the label after 80ms of fake-async time so
@@ -254,7 +112,7 @@ void main() {
   testWidgets(
     'fails with TestFailure when label never advances within timeout',
     (WidgetTester tester) async {
-      final controller = _NextTurnLabelController(
+      final controller = NextTurnLabelController(
         initialLabel: 'Next turn (1 / 1492)',
       );
       await _pumpHost(tester, controller);
@@ -281,7 +139,7 @@ void main() {
   testWidgets(
     'holds return until TurnResolutionProcessingDialog clears even if label already differs',
     (WidgetTester tester) async {
-      final controller = _NextTurnLabelController(
+      final controller = NextTurnLabelController(
         initialLabel: 'Next turn (2 / 1492)',
         initialShowProcessingDialog: true,
       );
@@ -333,7 +191,7 @@ void main() {
       'returns elapsed via post-loop final check when the while loop is '
       'skipped (Duration.zero) and the label already differs',
       (WidgetTester tester) async {
-        final controller = _NextTurnLabelController(
+        final controller = NextTurnLabelController(
           initialLabel: 'Next turn (2 / 1492)',
         );
         await _pumpHost(tester, controller);
@@ -367,7 +225,7 @@ void main() {
       'still fails with TestFailure when the label never advances and '
       'the loop is skipped (Duration.zero, additive contract)',
       (WidgetTester tester) async {
-        final controller = _NextTurnLabelController(
+        final controller = NextTurnLabelController(
           initialLabel: 'Next turn (1 / 1492)',
         );
         await _pumpHost(tester, controller);
@@ -402,40 +260,39 @@ void main() {
       },
     );
 
-    testWidgets(
-      'post-loop final check refuses to return while the '
-      'TurnResolutionProcessingDialog is still mounted',
-      (WidgetTester tester) async {
-        final controller = _NextTurnLabelController(
-          initialLabel: 'Next turn (2 / 1492)',
-          initialShowProcessingDialog: true,
-        );
-        await _pumpHost(tester, controller);
-        expect(find.byType(TurnResolutionProcessingDialog), findsOneWidget);
+    testWidgets('post-loop final check refuses to return while the '
+        'TurnResolutionProcessingDialog is still mounted', (
+      WidgetTester tester,
+    ) async {
+      final controller = NextTurnLabelController(
+        initialLabel: 'Next turn (2 / 1492)',
+        initialShowProcessingDialog: true,
+      );
+      await _pumpHost(tester, controller);
+      expect(find.byType(TurnResolutionProcessingDialog), findsOneWidget);
 
-        Object? caught;
-        try {
-          await e2eWaitForNextTurnLabelAdvance(
-            tester,
-            turnLabelBefore: 'Next turn (1 / 1492)',
-            timeout: Duration.zero,
-          );
-        } catch (e) {
-          caught = e;
-        }
-        expect(
-          caught,
-          isA<TestFailure>(),
-          reason:
-              'Even though the post-loop final check sees a different '
-              'label, it must observe `sawProcessingDialog=true` AND a '
-              'still-mounted dialog and therefore refuse to return — '
-              'matching the in-loop dialog gate so the post-loop fix '
-              'does not weaken the race contract pinned by the '
-              'sibling "holds return until …" test (#2336 AC5).',
+      Object? caught;
+      try {
+        await e2eWaitForNextTurnLabelAdvance(
+          tester,
+          turnLabelBefore: 'Next turn (1 / 1492)',
+          timeout: Duration.zero,
         );
-      },
-    );
+      } catch (e) {
+        caught = e;
+      }
+      expect(
+        caught,
+        isA<TestFailure>(),
+        reason:
+            'Even though the post-loop final check sees a different '
+            'label, it must observe `sawProcessingDialog=true` AND a '
+            'still-mounted dialog and therefore refuse to return — '
+            'matching the in-loop dialog gate so the post-loop fix '
+            'does not weaken the race contract pinned by the '
+            'sibling "holds return until …" test (#2336 AC5).',
+      );
+    });
   });
 
   testWidgets(
