+++ packages/colonizethis_app_e2e_support/test/e2e_wait_until_found_test.dart
@@ -22,84 +22,13 @@
 /// itself.
 library;
 
-import 'dart:async';
-
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
-// `dart:async` is imported for the `Timer` used by `_DelayedMountHost` to
-// flip its `_mounted` flag during the helper's adaptive pump loop without the
-// test driving extra pumps itself.
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
 
-/// Host that mounts a single keyed `TextButton` after an optional fake-async
-/// delay so the test can flip visibility while the helper polls.
-///
-/// `Timer` callbacks scheduled in [State.initState] fire when `tester.pump`
-/// advances fake-time past the registered duration, so the helper sees the
-/// newly mounted widget on a later iteration without the test calling
-/// `tester.pump` itself (which would deadlock against the helper's guarded
-/// pump loop).
-class _DelayedMountHost extends StatefulWidget {
-  const _DelayedMountHost({
-    required this.targetKey,
-    this.mountAfter,
-    this.startMounted = false,
-  });
-
-  /// Key the host will render once mounted; the test waits on
-  /// `find.byKey(targetKey)`.
-  final Key targetKey;
-
-  /// Fake-async delay before the host mounts [targetKey], or `null` to leave
-  /// the mounted flag untouched.
-  final Duration? mountAfter;
-
-  /// Whether [targetKey] is mounted on the first frame (pre-pump).
-  final bool startMounted;
-
-  @override
-  State<_DelayedMountHost> createState() => _DelayedMountHostState();
-}
-
-class _DelayedMountHostState extends State<_DelayedMountHost> {
-  late bool _mounted;
-  Timer? _mountTimer;
-
-  @override
-  void initState() {
-    super.initState();
-    _mounted = widget.startMounted;
-    final after = widget.mountAfter;
-    if (after != null) {
-      _mountTimer = Timer(after, () {
-        if (!mounted) return;
-        setState(() {
-          _mounted = true;
-        });
-      });
-    }
-  }
-
-  @override
-  void dispose() {
-    _mountTimer?.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (!_mounted) {
-      return const SizedBox.shrink();
-    }
-    return TextButton(
-      key: widget.targetKey,
-      onPressed: () {},
-      child: const Text('btn'),
-    );
-  }
-}
+import 'support/delayed_mount_harness.dart';
 
 Future<void> _pumpHost(
   WidgetTester tester, {
@@ -111,10 +40,14 @@ Future<void> _pumpHost(
     MaterialApp(
       home: Scaffold(
         body: Center(
-          child: _DelayedMountHost(
-            targetKey: targetKey,
+          child: DelayedMountHost(
             mountAfter: mountAfter,
             startMounted: startMounted,
+            child: TextButton(
+              key: targetKey,
+              onPressed: () {},
+              child: const Text('btn'),
+            ),
           ),
         ),
       ),
@@ -271,31 +204,30 @@ void main() {
     },
   );
 
-  testWidgets(
-    'accepts a custom phaseName and E2ePerfLog on the timeout path',
-    (WidgetTester tester) async {
-      const missingKey = Key('e2e_perf_timeout_btn');
-      await _pumpHost(tester, targetKey: missingKey);
-      final perf = E2ePerfLog('e2e_wait_until_found_test');
-      Object? caught;
-      try {
-        await e2eWaitUntilFound(
-          tester,
-          find.byKey(missingKey),
-          timeout: const Duration(milliseconds: 100),
-          perf: perf,
-          phaseName: 'pin_wait_until_found_perf_timeout',
-        );
-      } catch (e) {
-        caught = e;
-      }
-      expect(
-        caught,
-        isA<TestFailure>(),
-        reason:
-            'Passing a perf log on the timeout path must not suppress the '
-            'fail() invocation; perf is observability metadata only.',
+  testWidgets('accepts a custom phaseName and E2ePerfLog on the timeout path', (
+    WidgetTester tester,
+  ) async {
+    const missingKey = Key('e2e_perf_timeout_btn');
+    await _pumpHost(tester, targetKey: missingKey);
+    final perf = E2ePerfLog('e2e_wait_until_found_test');
+    Object? caught;
+    try {
+      await e2eWaitUntilFound(
+        tester,
+        find.byKey(missingKey),
+        timeout: const Duration(milliseconds: 100),
+        perf: perf,
+        phaseName: 'pin_wait_until_found_perf_timeout',
       );
-    },
-  );
+    } catch (e) {
+      caught = e;
+    }
+    expect(
+      caught,
+      isA<TestFailure>(),
+      reason:
+          'Passing a perf log on the timeout path must not suppress the '
+          'fail() invocation; perf is observability metadata only.',
+    );
+  });
 }
